backend/requirements.txt
asgiref==3.7.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0